requirements.txt
# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.1.0
cryptography>=41.0.4
django-ratelimit>=4.1.0
social-auth-app-django>=5.4.0

# API Documentation (Swagger)
drf-yasg>=1.21.7

# Celery & Redis
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Database
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
python-decouple>=3.8
requests>=2.31.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0
